+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in 11.8
 - HTML lower max Display Columns to 6 (was 10) in StudentFields.php & UserFields.php
 - Check is admin before saving Template in ReportCards.php & Transcripts.php
 - CSS remove forced large font size for Transcripts PDF in Transcripts.php
+- Add & get PROFILE_ID column in GetTeacher.fnc.php & User.fnc.php
 
 Changes in 11.7.3
 -----------------